feat(including bugs): minor changes that increases functionality (#1)
* Create .gitpod.yml

Need the VNC image for the GUI app

* Remove the gitpod file(not working) and change values in SPEC, main.py and add latest exe

* Add pycache folder to .gitignore

* Change path from local path to `.` and exe name

On Linux, if you enter `ls -a`, you get a directory known as `.`, which is the root dir. Works the same on macOS and Windows, so by adding that, it works universally.

* add root dir code

* Update main.yml

We need builds for each OS since TimerX could be cross-platform. Rather than starting a file and having Python installed, you can get an executable file for your OS.

* Update main.yml

Change py2exe to pyinstaller

* Create requirements.txt

* Add files via upload

* Linux icon file(test required)

* Add platform code

* Update main.py

Remove duplicate code.

* chore: Simplify Logic

* Update main.py

* Update .gitignore

Co-authored-by: coder-lg <76845820+coder-lg@users.noreply.github.com>
Co-authored-by: sumeshir26 <sumeshir26@protonmail.com>